First upload
Upload of

- Statistical analysis of steel design.py
This is an example of usage of simple statistics, and data manipulation using SQLite3 inside Python. With this program quick analysis of most stressed members in steel structure can be performed.

- Data Checker_3
This is an example of usage of Pandas library inside Python. With this program a quick analysis of groups within dataset can be performed.

Both files have their prepared example files inside the named directories